+++ src/main/java/ourtine/auth/service/AppleService.java
@@ -160,7 +160,6 @@ public class AppleService{
                     .participationRate(0)
                     .imageUrl(null)
                     .goal(null)
-                    .email(null)
                     .introduce(null)
                     .nickname(null)
                     .privacyAgreed(false)
