+++ src/main/java/ourtine/domain/User.java
@@ -30,7 +30,6 @@ public class User extends BaseEntity {
     @Column(nullable = false)
     private String nickname;
     private String introduce;
-    private String email;
     private String imageUrl;
     private String goal;
     @Column(nullable = false)
